% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.R
\name{calc.loglik}
\alias{calc.loglik}
\title{Calculate the log likelihood of the model}
\usage{
calc.loglik(Y, W, B, sigma2, tau2, a = 0.001, b = 0.001)
}
\arguments{
\item{Y}{is an NxP matrix}

\item{W}{is a PxD matrix}

\item{B}{is a DxN matrix}

\item{sigma2}{is a real number}

\item{tau2}{is a real number}

\item{a}{the first parameter of the inverse gamma distribution}

\item{b}{the second parameter of the inverse gamma distribution}
}
\value{
loglik the likelihood
}
\description{
Calculate the log likelihood of the model
}
\author{
Gabrielle Weinrott
}
\keyword{internal}

