% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.R
\name{coda.obj}
\alias{coda.obj}
\title{Convert a STAN objet to MCMC list}
\usage{
coda.obj(stanfit)
}
\arguments{
\item{stanfit}{a STAN object}
}
\value{
codafit an mcmc.list
}
\description{
Convert a STAN objet to MCMC list
}
\examples{
data(stanfit) # output of modelFit or main.modelFit
coda.fit <- coda.obj(stanfit)
head(coda.fit)

}
\author{
Gabrielle Weinrott
}

