% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{interval.pca.Deville}
\alias{interval.pca.Deville}
\title{Perform Interval-PCA using Deville's method}
\usage{
interval.pca.Deville(X, t, t.range, breaks)
}
\arguments{
\item{X}{a data matrix}

\item{t}{a matrix of observation times corresponding to X}

\item{t.range}{the range of observation times in vector form (ex. t.range = c(0, 1000))}

\item{breaks}{integer number of histogram windows}
}
\value{
scores

scores.min

scores.max
}
\description{
Perform Interval-PCA using Deville's method
}
\author{
Gabrielle Weinrott
}
\references{
JC Deville, "Methodes statisiques et numeriques de l'analyse harmonique", Annales de l'INSEE, 1974.
}

