% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{main.modelFit}
\alias{main.modelFit}
\title{Main function to fit the Bayesian Latent Factor model}
\usage{
main.modelFit(params, out.file, verbose = 1, obj.version)
}
\arguments{
\item{params}{a list of parameters for the modelFit function, see modelFit help for details}

\item{out.file}{the path to the file that will be saved}

\item{verbose}{binary variable (default = 1)}

\item{obj.version}{output version (eg. "1.1.plt")}
}
\value{
nothing
}
\description{
Main function to fit the Bayesian Latent Factor model
}
\author{
Gabrielle Weinrott
}
\keyword{internal}

