% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.R
\name{postdens}
\alias{postdens}
\title{Calculate the unnormalized posterior density of the model}
\usage{
postdens(mcmc.output, Y, D, chain = 1)
}
\arguments{
\item{mcmc.output}{an mcmc list as produced by clean.mcmc}

\item{Y}{the data matrix}

\item{D}{the number of latent factors}

\item{chain}{the chain to plot (default = 1)}
}
\value{
post a vector containing the posterior density at each iteration##' @examples
}
\description{
Calculate the unnormalized posterior density of the model
}
\examples{
data("toydata")
data("stanfit")
dens <- postdens(coda.obj(stanfit), Y = toydata$Y.simul$Y, D = 2, chain = 1)
hist(dens)

}
\author{
Gabrielle Weinrott
}

