% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\docType{data}
\name{stanfit}
\alias{stanfit}
\title{A stanfit object fitted to the toydata}
\format{
A large stanfit object
}
\usage{
stanfit
}
\description{
A stanfit object fitted to the toydata
}
\keyword{datasets}
